refactoring - optimization
wow - can't believe how hard I resisted simply adding other references to the Reservation joiner model. May this serve as a good lesson about not being too rigid with your coding practices. adding a few extra strings to the reservation object is not distracting or convoluted in anyway, and is self explanatory because the reservation model is already so simple. day, time, and student's name is awesome info to have with one fetch! no need more multiple fetches on the fornt end or complex back end controller logic.